Use readlink to canonicalize path to script
readlink without `-f` will return the relative value of the symlink and
realpath will fail for values such as `../scm/repo/script.sh` when this
is in `~/bin` and the script is called via the path. readlink will
properly canonicalize the path when `-f` is specified, therefore just
use it if available. readlink on MacOS does not support it.